requirements/dev.txt
django-debug-toolbar==1.7

